feat: add project file browser, initialization for BIDS projects, and open project controller
- Implemented a file browser for selecting existing project files with a modal interface.
- Created a controller for initializing PRISM projects from BIDS datasets, including user feedback and error handling.
- Developed an open project controller to manage loading and validating projects, with UI updates for project state.
- Added path picker functionality for selecting project and dataset locations.
- Introduced shared macros for rendering UI components consistently across templates.
- Enhanced home page with informative sections about PRISM Studio features and project structure.
- Added tests for rendering shared components and validating project paths.
