+++ app/templates/includes/home/highlights.html
@@ -1,0 +1,78 @@
+{% from "includes/ui/macros.html" import section_card %}
+
+{% call section_card('Why Researchers Use PRISM Studio', 'fas fa-star', 'text-success', 'feature-card home-feature-card mb-4') %}
+<div class="row g-4">
+    <div class="col-md-6">
+        <div class="mb-3">
+            <h6 class="mb-2 home-highlight-title">
+                <i class="fas fa-file-import me-2 home-highlight-icon"></i>
+                Convert messy source data
+            </h6>
+            <p class="text-muted small mb-0">
+                Import Excel, CSV, TSV, and LimeSurvey data into a structured PRISM dataset without
+                hand-building folders or filenames.
+            </p>
+        </div>
+    </div>
+    <div class="col-md-6">
+        <div class="mb-3">
+            <h6 class="mb-2 home-highlight-title">
+                <i class="fas fa-calculator me-2 home-highlight-icon"></i>
+                Score questionnaires automatically
+            </h6>
+            <p class="text-muted small mb-0">
+                Build scoring recipes for surveys and compute derived scores without maintaining fragile
+                spreadsheet formulas.
+            </p>
+        </div>
+    </div>
+    <div class="col-md-6">
+        <div class="mb-3">
+            <h6 class="mb-2 home-highlight-title">
+                <i class="fas fa-chart-line me-2 home-highlight-icon"></i>
+                Export for analysis
+            </h6>
+            <p class="text-muted small mb-0">
+                Generate analysis-ready exports such as SPSS, CSV, TSV, labels, and codebooks from the
+                same project.
+            </p>
+        </div>
+    </div>
+    <div class="col-md-6">
+        <div class="mb-3">
+            <h6 class="mb-2 home-highlight-title">
+                <i class="fas fa-puzzle-piece me-2 home-highlight-icon"></i>
+                Stay compatible with BIDS apps
+            </h6>
+            <p class="text-muted small mb-0">
+                PRISM adds surveys and other study metadata without giving up standard BIDS validators
+                and downstream tools.
+            </p>
+        </div>
+    </div>
+    <div class="col-md-6">
+        <div class="mb-3">
+            <h6 class="mb-2 home-highlight-title">
+                <i class="fas fa-layer-group me-2 home-highlight-icon"></i>
+                Work across multiple modalities
+            </h6>
+            <p class="text-muted small mb-0">
+                Organize surveys, biometrics, eyetracking, physio, and environment data in one coherent
+                project structure.
+            </p>
+        </div>
+    </div>
+    <div class="col-md-6">
+        <div class="mb-3">
+            <h6 class="mb-2 home-highlight-title">
+                <i class="fas fa-user-shield me-2 home-highlight-icon"></i>
+                Clean up participant metadata
+            </h6>
+            <p class="text-muted small mb-0">
+                Remap participant IDs, standardize demographics, and prepare ontology-friendly metadata
+                while keeping processing on your machine.
+            </p>
+        </div>
+    </div>
+</div>
+{% endcall %}
