+++ app/templates/includes/home/structure.html
@@ -1,0 +1,32 @@
+{% from "includes/ui/macros.html" import section_card %}
+
+{% call section_card('One Structure For Multimodal Studies', 'fas fa-sitemap', 'text-success', 'feature-card home-structure-card mb-4') %}
+<div class="row g-4">
+    <div class="col-md-6">
+        <pre class="bg-light p-3 rounded border-success"><code>dataset/
+├── dataset_description.json
+├── participants.tsv
+├── sub-01/
+│   └── ses-001/
+│       ├── survey/
+│       ├── biometrics/
+│       └── environment/
+└── sub-02/
+    └── ...</code></pre>
+    </div>
+    <div class="col-md-6">
+        <h6 class="text-primary mb-2">What PRISM adds to a BIDS-style project</h6>
+        <p class="text-muted small mb-3">
+            Keep a familiar BIDS layout while adding the extra structure needed for surveys,
+            participant-level assessments, and richer behavioral studies.
+        </p>
+        <ul class="list-unstyled text-muted small mb-0">
+            <li class="mb-2"><i class="fas fa-chart-line text-success me-2"></i><code>survey/</code> - Questionnaires, scoring inputs, and behavioral responses</li>
+            <li class="mb-2"><i class="fas fa-heartbeat text-success me-2"></i><code>biometrics/</code> - Fitness, strength, balance, and other assessments</li>
+            <li class="mb-2"><i class="fas fa-cloud-sun text-success me-2"></i><code>environment/</code> - Privacy-safe contextual enrichment data</li>
+            <li class="mb-2"><i class="fas fa-eye text-success me-2"></i><code>eyetracking/</code> - Eye tracking workflows in the same study layout</li>
+            <li><i class="fas fa-wave-square text-success me-2"></i><code>physiological/</code> - Standard BIDS physio alongside PRISM extensions</li>
+        </ul>
+    </div>
+</div>
+{% endcall %}
