+++ app/templates/includes/projects/init_bids_form.html
@@ -1,3 +1,6 @@
+    {% from "includes/ui/macros.html" import path_picker %}
+    {% from "includes/projects/macros.html" import project_flow_strip %}
+
     <!-- Init PRISM on existing BIDS dataset -->
     <div class="form-section" id="section-init-bids">
         <div class="card shadow-sm">
@@ -5,20 +8,11 @@
                 <h5 class="mb-0"><i class="fas fa-layer-group text-warning me-2"></i>Init PRISM on Existing BIDS Dataset</h5>
             </div>
             <div class="card-body">
-                <div class="project-flow-strip project-flow-strip--init mb-3" id="initBidsFlowStrip">
-                    <div class="project-flow-strip__item">
-                        <span class="project-flow-strip__label">Use when</span>
-                        <strong>Your dataset already has a valid BIDS root</strong>
-                    </div>
-                    <div class="project-flow-strip__item">
-                        <span class="project-flow-strip__label">PRISM adds</span>
-                        <strong>Project metadata, helper files, and standard support folders</strong>
-                    </div>
-                    <div class="project-flow-strip__item">
-                        <span class="project-flow-strip__label">Keeps intact</span>
-                        <strong>Existing subject data, scans, and BIDS sidecars</strong>
-                    </div>
-                </div>
+                {{ project_flow_strip('project-flow-strip--init', 'initBidsFlowStrip', [
+                    ('Use when', 'Your dataset already has a valid BIDS root'),
+                    ('PRISM adds', 'Project metadata, helper files, and standard support folders'),
+                    ('Keeps intact', 'Existing subject data, scans, and BIDS sidecars')
+                ]) }}
                 <div class="alert alert-info mb-3">
                     <i class="fas fa-info-circle me-2"></i>
                     Point to your existing BIDS root (the folder containing <code>dataset_description.json</code>).
@@ -29,15 +23,9 @@
                 <form id="initBidsForm">
                     <div class="row g-3">
                         <div class="col-md-8">
-                            <label for="initBidsPath" class="form-label fw-bold">BIDS Dataset Root <span class="badge bg-danger">REQUIRED</span></label>
-                            <div class="input-group">
-                                <input type="text" class="form-control" id="initBidsPath"
-                                       placeholder="/path/to/your-bids-dataset" required>
-                                <button class="btn btn-outline-secondary" type="button" id="browseInitBidsPath">
-                                    <i class="fas fa-folder-open"></i> Browse
-                                </button>
-                            </div>
-                            <small class="text-muted">Must contain <code>dataset_description.json</code></small>
+                            {% call path_picker('BIDS Dataset Root', 'initBidsPath', 'browseInitBidsPath', '/path/to/your-bids-dataset', 'Must contain dataset_description.json', 'Browse', '', '', true, '', '', 'form-label fw-bold') %}
+                            <span class="badge bg-danger">REQUIRED</span>
+                            {% endcall %}
                         </div>
                         <div class="col-md-4">
                             <label for="initBidsName" class="form-label fw-bold">Display Name <span class="badge bg-secondary">optional</span></label>
