+++ app/templates/index.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
-{% from "includes/ui/macros.html" import page_header %}
+{% from "includes/ui/macros.html" import page_header, section_card %}
 
 {% block title %}PRISM Studio - Dataset Validation{% endblock %}
 
@@ -47,12 +47,7 @@
         ) }}
 
         <!-- Main Upload Card -->
-        <div class="card feature-card mb-4 border-success">
-            <div class="card-body p-5">
-                <h3 class="card-title text-start mb-4">
-                    <i class="fas fa-check-circle text-success me-2"></i>
-                    Dataset Validation
-                </h3>
+        {% call section_card('Dataset Validation', 'fas fa-check-circle', 'text-success', 'feature-card mb-4 border-success', 'p-5', 'card-title text-start mb-4') %}
 
                 <p class="text-start text-muted mb-4">
                     The PRISM Validator ensures your dataset follows the PRISM structure and BIDS standards.
@@ -152,8 +147,7 @@
                                     </label>
                                     <div>
                                         <div class="form-check mb-2">
-                                            <input class="form-check-input advanced-option" type="radio" name="validation_mode" id="mode_both" value="both" checked disabled>
-                                            <label class="form-check-label fw-bold" for="mode_both">Full Validation (PRISM + BIDS)</label>
+                {% endcall %}
                                         </div>
                                         <div class="form-check mb-2">
                                             <input class="form-check-input advanced-option" type="radio" name="validation_mode" id="mode_prism" value="prism" disabled>
