+++ tests/test_navbar_project_event_wiring.py
@@ -18,6 +18,16 @@ class TestNavbarProjectEventWiring(unittest.TestCase):
         self.assertIn("setFileManagementState(Boolean(nextPath));", content)
         self.assertIn("setDerivativesState(Boolean(nextPath));", content)
 
+    def test_navbar_recent_project_loader_uses_credentialed_fallback_fetch(self):
+        content = BASE_TEMPLATE.read_text(encoding="utf-8")
+
+        self.assertIn("function normalizeNavbarFetchOptions(url, options = {})", content)
+        self.assertIn("credentials: 'include'", content)
+        self.assertIn(
+            "return fetch(fallbackUrl, normalizeNavbarFetchOptions(fallbackUrl, options));",
+            content,
+        )
+
 
 if __name__ == "__main__":
     unittest.main()
